backend/requirements.txt
flask
flask-socketio
eventlet
python-can
joblib
numpy
pandas
influxdb-client